+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.1.0+1
+# Changelog
+
+## <a id="v0.1.1"></a>[v0.1.1] - 2024-05-30
+
+### Added
+- more unit tests
+- new util functions
+- nullable string extensions
+
+### Fixed
+- collections unit test file name
+
+## <a id="v0.1.0+1"></a>[v0.1.0+1] - 2024-05-29
 
 - Initial release.